requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8
openai==1.3.5
pymupdf==1.23.7
faiss-cpu==1.7.4
websockets==12.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
tiktoken==0.5.1
numpy==1.26.2
pandas==2.1.3
httpx==0.23.3
scikit-learn==1.7.0
scipy==1.16.0